报警表数据清理，未测试

+++ src/test/java/com/zdzc/dataClear/DataTest.java
@@ -24,6 +24,6 @@ public class DataTest {
 
     @Test
     public void test3(){
-        System.out.println(Long.MAX_VALUE);
+        System.out.println(Integer.MAX_VALUE);
     }
 }
